Time for aures to spin.

+++ src/main/java/by/jackraidenph/dragonsurvival/magic/common/AbilityAnimation.java
@@ -3,10 +3,10 @@ package by.jackraidenph.dragonsurvival.magic.common;
 public class AbilityAnimation
 {
 	public String animationKey;
-	public int duration;
+	public double duration;
 	public boolean locksNeck;
 	
-	public AbilityAnimation(String animationKey, int duration, boolean locksNeck)
+	public AbilityAnimation(String animationKey, double duration, boolean locksNeck)
 	{
 		this(animationKey, locksNeck);
 		this.duration = duration;
